Update and split Solver into different modules.
